subset model to overlap with tables before checking keys
'tables' input  may not contain all tables in the model, and primary/
foreign key checks would fail because they were trying to check
keys in tables that did not exist. Subset before checking. This can
be done with an argument for primary keys, but for foreign keys the
entire model must be subset.