add_auto_columns has the option to do nothing if cols are missing
We want to be able to run add_auto_columns before running the
data model report. Therefore it must have the option to continue
without error even if some columns are missing. If not all columns
are present and error_on_missing is false, a warning will be
issued and the column will not be added.