add support for conditional requirements
Conditional requirements can be specifed for both tables (e.g.,
table t2 is required if table t1 is present), and columns (e.g.,
column x is required if any values of column y are TRUE) These
conditions are stored in the data model object, and parsed by the
check function into required and optional columns. The major
change to previous behavior is that the "required" attribute is a
vector of table or column names rather than a logical vector.