Fixes #11. AnvilDataModels is now compatible with dm 1.0
In dm v1.0, all foreign keys in the dm object must have unique
values. This is not true for keys in set tables, which are not
unique until they have been imported to AnVIL. Add an additional
check for values in set keys, but do not include them in the call
to dm_examine_constraints.
Another change introduced in dm v1.0 was that unique keys are
automatically detected and include in the output of
dm_examine_constraints, even if not explicitly declared as
primary keys. Foreign key checks need to filter on keys with
kind = 'FK'.