allow foreign keys with multi-value delimiters
The dm foreign key check will not work if the key in the child
table is separated with multi-value delimiters. Check this as a
special case, and add any problems to set_key_problems if found
(arguably this is a kind of set and it makes it easier to use
existing code to represent a different type of foriegn key
problem).